Abstract client common classes to basic module.
