+++ client-basic/src/main/java/com/alibaba/nacos/client/address/AbstractServerListProvider.java
@@ -19,7 +19,7 @@ package com.alibaba.nacos.client.address;
 import com.alibaba.nacos.api.PropertyKeyConst;
 import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.client.env.NacosClientProperties;
-import com.alibaba.nacos.client.utils.ParamUtil;
+import com.alibaba.nacos.client.utils.ClientBasicParamUtil;
 import com.alibaba.nacos.common.http.client.NacosRestTemplate;
 import com.alibaba.nacos.common.utils.StringUtils;
 
@@ -32,7 +32,7 @@ import java.util.List;
  */
 public abstract class AbstractServerListProvider implements ServerListProvider {
     
-    protected String contextPath = ParamUtil.getDefaultContextPath();
+    protected String contextPath = ClientBasicParamUtil.getDefaultContextPath();
     
     protected String namespace = "";
