+++ client-basic/src/main/java/com/alibaba/nacos/client/address/ServerListProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2024 Alibaba Group Holding Ltd.
+ * Copyright 1999-2025 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@ package com.alibaba.nacos.client.address;
 
 import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.client.env.NacosClientProperties;
-import com.alibaba.nacos.client.utils.ParamUtil;
+import com.alibaba.nacos.client.utils.ClientBasicParamUtil;
 import com.alibaba.nacos.common.http.client.NacosRestTemplate;
 import com.alibaba.nacos.common.lifecycle.Closeable;
 
@@ -66,7 +66,7 @@ public interface ServerListProvider extends Closeable {
      * @return context path
      */
     default String getContextPath() {
-        return ParamUtil.getDefaultContextPath();
+        return ClientBasicParamUtil.getDefaultContextPath();
     }
     
     /**
