+++ client-basic/src/main/java/com/alibaba/nacos/client/utils/AppNameUtils.java
