+++ client-basic/src/test/java/com/alibaba/nacos/client/utils/AppNameUtilsTest.java
