+++ client/src/test/java/com/alibaba/nacos/client/naming/utils/InitUtilsTest.java
@@ -146,91 +146,4 @@ class InitUtilsTest {
         assertEquals("/nacos/v1/ns", UtilAndComs.nacosUrlBase);
         assertEquals("/nacos/v1/ns/instance", UtilAndComs.nacosUrlInstance);
     }
-    
-    @Test
-    void testInitEndpointForNullProperties() {
-        assertEquals("", InitUtils.initEndpoint(null));
-    }
-    
-    @Test
-    void testInitEndpointFromDefaultWithoutCloudParsing() {
-        System.setProperty(SystemPropertyKeyConst.IS_USE_ENDPOINT_PARSING_RULE, "false");
-        final NacosClientProperties properties = NacosClientProperties.PROTOTYPE.derive();
-        String actual = InitUtils.initEndpoint(properties);
-        assertEquals("", actual);
-    }
-    
-    @Test
-    void testInitEndpointFromPropertiesWithoutCloudParsing() {
-        System.setProperty(SystemPropertyKeyConst.IS_USE_ENDPOINT_PARSING_RULE, "false");
-        final NacosClientProperties properties = NacosClientProperties.PROTOTYPE.derive();
-        String endpoint = "1.1.1.1";
-        String endpointPort = "1234";
-        properties.setProperty(PropertyKeyConst.ENDPOINT, endpoint);
-        properties.setProperty(PropertyKeyConst.ENDPOINT_PORT, endpointPort);
-        String actual = InitUtils.initEndpoint(properties);
-        assertEquals(endpoint + ":" + endpointPort, actual);
-    }
-    
-    @Test
-    void testInitEndpointFromAliwareWithoutCloudParsing() {
-        String endpoint = "aliware_endpoint";
-        String endpointPort = "1234";
-        System.setProperty(SystemPropertyKeyConst.IS_USE_ENDPOINT_PARSING_RULE, "false");
-        System.setProperty(PropertyKeyConst.SystemEnv.ALIBABA_ALIWARE_ENDPOINT_URL, endpoint);
-        System.setProperty(PropertyKeyConst.SystemEnv.ALIBABA_ALIWARE_ENDPOINT_PORT, endpointPort);
-        final NacosClientProperties properties = NacosClientProperties.PROTOTYPE.derive();
-        properties.setProperty(PropertyKeyConst.ENDPOINT_PORT, endpointPort + "1");
-        String actual = InitUtils.initEndpoint(properties);
-        assertEquals("", actual);
-    }
-    
-    @Test
-    void testInitEndpointFromDefaultWithCloudParsing() {
-        System.setProperty(SystemPropertyKeyConst.IS_USE_ENDPOINT_PARSING_RULE, "true");
-        final NacosClientProperties properties = NacosClientProperties.PROTOTYPE.derive();
-        String actual = InitUtils.initEndpoint(properties);
-        assertEquals("", actual);
-    }
-    
-    @Test
-    void testInitEndpointFromPropertiesWithCloudParsing() {
-        System.setProperty(SystemPropertyKeyConst.IS_USE_ENDPOINT_PARSING_RULE, "true");
-        final NacosClientProperties properties = NacosClientProperties.PROTOTYPE.derive();
-        String endpoint = "1.1.1.1";
-        String endpointPort = "1234";
-        properties.setProperty(PropertyKeyConst.ENDPOINT, endpoint);
-        properties.setProperty(PropertyKeyConst.ENDPOINT_PORT, endpointPort);
-        String actual = InitUtils.initEndpoint(properties);
-        assertEquals(endpoint + ":" + endpointPort, actual);
-    }
-    
-    @Test
-    void testInitEndpointFromAliwareWithCloudParsing() {
-        String endpoint = "aliware_endpoint";
-        String endpointPort = "1234";
-        System.setProperty(SystemPropertyKeyConst.IS_USE_ENDPOINT_PARSING_RULE, "true");
-        System.setProperty(PropertyKeyConst.SystemEnv.ALIBABA_ALIWARE_ENDPOINT_URL, endpoint);
-        System.setProperty(PropertyKeyConst.SystemEnv.ALIBABA_ALIWARE_ENDPOINT_PORT, endpointPort);
-        final NacosClientProperties properties = NacosClientProperties.PROTOTYPE.derive();
-        properties.setProperty(PropertyKeyConst.ENDPOINT_PORT, endpointPort + "1");
-        String actual = InitUtils.initEndpoint(properties);
-        assertEquals(endpoint + ":" + endpointPort, actual);
-    }
-    
-    @Test
-    void testInitEndpointAns() {
-        try {
-            System.setProperty(PropertyKeyConst.IS_USE_ENDPOINT_PARSING_RULE, "true");
-            final NacosClientProperties properties = NacosClientProperties.PROTOTYPE.derive();
-            String endpoint = "${key:test.com}";
-            properties.setProperty(PropertyKeyConst.ENDPOINT, endpoint);
-            String actual = InitUtils.initEndpoint(properties);
-            //defaultEndpointPort is  "8080";
-            assertEquals("test.com:8080", actual);
-        } finally {
-            System.clearProperty(PropertyKeyConst.IS_USE_ENDPOINT_PARSING_RULE);
-        }
-    }
-    
 }
