+++ common/src/main/java/com/alibaba/nacos/common/utils/CollectionUtils.java
@@ -22,6 +22,7 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Enumeration;
+import java.util.HashMap;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -35,6 +36,11 @@ import java.util.Set;
  * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
  */
 public final class CollectionUtils {
+    
+    /**
+     * Constant to avoid repeated object creation.
+     */
+    private static final Integer INTEGER_ONE = 1;
 
     private CollectionUtils() {
     }
@@ -305,34 +311,6 @@ public final class CollectionUtils {
         throw new IllegalArgumentException(buildExceptionMessage(iterator, first));
     }
     
-    /**
-     * check list is equal.
-     *
-     * @param firstList  first list.
-     * @param secondList second list.
-     * @return
-     */
-    public static boolean isListEqual(List<String> firstList, List<String> secondList) {
-        if (firstList == null && secondList == null) {
-            return true;
-        }
-        if (firstList == null || secondList == null) {
-            return false;
-        }
-        
-        if (firstList == secondList) {
-            return true;
-        }
-        
-        if (firstList.size() != secondList.size()) {
-            return false;
-        }
-        
-        boolean flag1 = firstList.containsAll(secondList);
-        boolean flag2 = secondList.containsAll(firstList);
-        return flag1 && flag2;
-    }
-    
     @SuppressWarnings("PMD.UndefineMagicConstantRule")
     private static <T> String buildExceptionMessage(Iterator<T> iterator, T first) {
         StringBuilder msg = new StringBuilder();
@@ -357,5 +335,68 @@ public final class CollectionUtils {
      */
     public static boolean isMapEmpty(Map<?, ?> map) {
         return (map == null || map.isEmpty());
+    }
+    
+    /**
+     * Returns a {@link Map} mapping each unique element in the given {@link Collection} to an {@link Integer}
+     * representing the number of occurrences of that element in the {@link Collection}.
+     *
+     * <p>Only those elements present in the collection will appear as keys in the map.
+     *
+     * @param coll the collection to get the cardinality map for, must not be null
+     * @return the populated cardinality map
+     */
+    public static Map getCardinalityMap(final Collection coll) {
+        Map count = new HashMap(coll.size());
+        for (Iterator it = coll.iterator(); it.hasNext(); ) {
+            Object obj = it.next();
+            Integer c = (Integer) (count.get(obj));
+            if (c == null) {
+                count.put(obj, INTEGER_ONE);
+            } else {
+                count.put(obj, c + 1);
+            }
+        }
+        return count;
+    }
+    
+    /**
+     * Returns <tt>true</tt> iff the given {@link Collection}s contain exactly the same elements with exactly the same
+     * cardinalities.
+     *
+     * <p>That is, iff the cardinality of <i>e</i> in <i>a</i> is equal to the cardinality of <i>e</i> in <i>b</i>, for
+     * each element <i>e</i> in <i>a</i> or <i>b</i>.
+     *
+     * @param a the first collection, must not be null
+     * @param b the second collection, must not be null
+     * @return <code>true</code> iff the collections contain the same elements with the same cardinalities.
+     */
+    public static boolean isEqualCollection(final Collection a, final Collection b) {
+        if (a.size() != b.size()) {
+            return false;
+        } else {
+            Map mapa = getCardinalityMap(a);
+            Map mapb = getCardinalityMap(b);
+            if (mapa.size() != mapb.size()) {
+                return false;
+            } else {
+                Iterator it = mapa.keySet().iterator();
+                while (it.hasNext()) {
+                    Object obj = it.next();
+                    if (getFreq(obj, mapa) != getFreq(obj, mapb)) {
+                        return false;
+                    }
+                }
+                return true;
+            }
+        }
+    }
+    
+    private static int getFreq(final Object obj, final Map freqMap) {
+        Integer count = (Integer) freqMap.get(obj);
+        if (count != null) {
+            return count.intValue();
+        }
+        return 0;
     }
 }
