+++ pom.xml
@@ -641,6 +641,7 @@
         <module>server</module>
         <module>lock</module>
         <module>maintainer-client</module>
+        <module>client-basic</module>
     </modules>
     
     <!-- Default dependencies in all subprojects -->
@@ -855,6 +856,11 @@
             <dependency>
                 <groupId>${project.groupId}</groupId>
                 <artifactId>nacos-sys</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.alibaba.nacos</groupId>
+                <artifactId>nacos-client-basic</artifactId>
                 <version>${project.version}</version>
             </dependency>
